requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
shapely>=2.0
alphashape>=1.3.1
scipy>=1.10
matplotlib>=3.7
